added progressBar component and stickers carousel
