+++ src/screens/mainScreen/Album/AlbumNavigator.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import Album from '../Album'
-import AlbumPage from '../AlbumPage'
+import Album from './Album'
+import AlbumPage from './AlbumPage'
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
 
 const Stack = createNativeStackNavigator();
